+++ CHANGELOG.md
@@ -58,6 +58,7 @@ Release 5.9.0 (unreleased):
    - In `WalletService` deprecate constructor parameters `requestEncryption`, `decryptionKeyMaterial`, `supportedJweAlgorithm`, `supportedJweEncryptionAlgorithm`
    - In `WalletService` introduce constructor parameter `encryptionService` which handles credential request encryption and credential response decryption
    - In `WalletService` add method `parseCredentialResponse` to transform the received credential response from the issuer into `StoreCredentialInput`
+   - In `WalletService` deprecate method `createCredentialRequest` and replace it with `createCredential` to handle encryption
 
 Release 5.8.0:
  - Refactor `AuthorizationServiceStrategy`